+++ tinyemu-core/riscv_cpu_priv.h
@@ -162,6 +162,7 @@ struct RISCVCPUState {
     void *time_opaque;
     BOOL power_down_flag;
     BOOL host_attention;
+    BOOL timer_attention;
     int pending_exception; /* used during MMU exception handling */
     target_ulong pending_tval;